First attempt to cleanup worker code. - Moved replayer, register and encoding to their own files. - Moded some util functions to common. - No functional change.
