+++ internal/internal_worker.go
@@ -24,16 +24,12 @@ package internal
 // All code in this file is private to the package.
 
 import (
-	"bytes"
 	"context"
 	"crypto/md5"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"math"
 	"os"
 	"reflect"
 	"runtime"
@@ -42,18 +38,14 @@ import (
 	"sync"
 	"time"
 
-	"github.com/apache/thrift/lib/go/thrift"
-	"github.com/golang/mock/gomock"
 	"github.com/opentracing/opentracing-go"
 	"github.com/pborman/uuid"
 	"github.com/uber-go/tally"
 	"go.uber.org/cadence/.gen/go/cadence/workflowserviceclient"
-	"go.uber.org/cadence/.gen/go/cadence/workflowservicetest"
 	"go.uber.org/cadence/.gen/go/shared"
-	"go.uber.org/cadence/internal/common"
 	"go.uber.org/cadence/internal/common/backoff"
 	"go.uber.org/cadence/internal/common/metrics"
-	"go.uber.org/cadence/internal/common/serializer"
+	"go.uber.org/cadence/internal/common/util"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
@@ -532,270 +524,6 @@ func (aw *activityWorker) Stop() {
 	aw.worker.Stop()
 }
 
-type registry struct {
-	sync.Mutex
-	workflowFuncMap  map[string]interface{}
-	workflowAliasMap map[string]string
-	activityFuncMap  map[string]activity
-	activityAliasMap map[string]string
-	next             *registry // Allows to chain registries
-}
-
-func (r *registry) RegisterWorkflow(af interface{}) {
-	r.RegisterWorkflowWithOptions(af, RegisterWorkflowOptions{})
-}
-
-func (r *registry) RegisterWorkflowWithOptions(
-	wf interface{},
-	options RegisterWorkflowOptions,
-) {
-	// Validate that it is a function
-	fnType := reflect.TypeOf(wf)
-	if err := validateFnFormat(fnType, true); err != nil {
-		panic(err)
-	}
-	fnName := getFunctionName(wf)
-	alias := options.Name
-	registerName := fnName
-	if len(alias) > 0 {
-		registerName = alias
-	}
-
-	r.Lock()
-	defer r.Unlock()
-
-	if !options.DisableAlreadyRegisteredCheck {
-		if _, ok := r.workflowFuncMap[registerName]; ok {
-			panic(fmt.Sprintf("workflow name \"%v\" is already registered", registerName))
-		}
-	}
-	r.workflowFuncMap[registerName] = wf
-	if len(alias) > 0 {
-		r.workflowAliasMap[fnName] = alias
-	}
-}
-
-func (r *registry) RegisterActivity(af interface{}) {
-	r.RegisterActivityWithOptions(af, RegisterActivityOptions{})
-}
-
-func (r *registry) RegisterActivityWithOptions(af interface{}, options RegisterActivityOptions) {
-	fnType := reflect.TypeOf(af)
-	var err error
-	if fnType.Kind() == reflect.Ptr && fnType.Elem().Kind() == reflect.Struct {
-		err = r.registerActivityStruct(af, options)
-	} else {
-		err = r.registerActivityFunction(af, options)
-	}
-
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (r *registry) registerActivityFunction(af interface{}, options RegisterActivityOptions) error {
-	fnType := reflect.TypeOf(af)
-	if err := validateFnFormat(fnType, false); err != nil {
-		return fmt.Errorf("failed to register activity method: %v", err)
-	}
-
-	fnName := getFunctionName(af)
-	alias := options.Name
-	registerName := fnName
-	if len(alias) > 0 {
-		registerName = alias
-	}
-
-	r.Lock()
-	defer r.Unlock()
-
-	if !options.DisableAlreadyRegisteredCheck {
-		if _, ok := r.activityFuncMap[registerName]; ok {
-			return fmt.Errorf("activity type \"%v\" is already registered", registerName)
-		}
-	}
-	r.activityFuncMap[registerName] = &activityExecutor{registerName, af}
-	if len(alias) > 0 {
-		r.activityAliasMap[fnName] = alias
-	}
-
-	return nil
-}
-
-func (r *registry) registerActivityStruct(aStruct interface{}, options RegisterActivityOptions) error {
-	r.Lock()
-	defer r.Unlock()
-
-	structValue := reflect.ValueOf(aStruct)
-	structType := structValue.Type()
-	count := 0
-	for i := 0; i < structValue.NumMethod(); i++ {
-		methodValue := structValue.Method(i)
-		method := structType.Method(i)
-		// skip private method
-		if method.PkgPath != "" {
-			continue
-		}
-		name := method.Name
-		if err := validateFnFormat(method.Type, false); err != nil {
-			return fmt.Errorf("failed to register activity method %v of %v: %e", name, structType.Name(), err)
-		}
-		prefix := options.Name
-		registerName := name
-		if len(prefix) == 0 {
-			prefix = structType.Elem().Name() + "_"
-		}
-		registerName = prefix + name
-		if !options.DisableAlreadyRegisteredCheck {
-			if _, ok := r.getActivityNoLock(registerName); ok {
-				return fmt.Errorf("activity type \"%v\" is already registered", registerName)
-			}
-		}
-		r.activityFuncMap[registerName] = &activityExecutor{registerName, methodValue.Interface()}
-		count++
-	}
-
-	if count == 0 {
-		return fmt.Errorf("no activities (public methods) found in %v structure", structType.Name())
-	}
-
-	return nil
-}
-
-func (r *registry) getWorkflowAlias(fnName string) (string, bool) {
-	r.Lock() // do not defer for Unlock to call next.getWorkflowAlias without lock
-	alias, ok := r.workflowAliasMap[fnName]
-	if !ok && r.next != nil {
-		r.Unlock()
-		return r.next.getWorkflowAlias(fnName)
-	}
-	r.Unlock()
-	return alias, ok
-}
-
-func (r *registry) getWorkflowFn(fnName string) (interface{}, bool) {
-	r.Lock() // do not defer for Unlock to call next.getWorkflowFn without lock
-	fn, ok := r.workflowFuncMap[fnName]
-	if !ok && r.next != nil {
-		r.Unlock()
-		return r.next.getWorkflowFn(fnName)
-	}
-	r.Unlock()
-	return fn, ok
-}
-
-func (r *registry) getRegisteredWorkflowTypes() []string {
-	r.Lock() // do not defer for Unlock to call next.getRegisteredWorkflowTypes without lock
-	var result []string
-	for t := range r.workflowFuncMap {
-		result = append(result, t)
-	}
-	r.Unlock()
-	if r.next != nil {
-		nextTypes := r.next.getRegisteredWorkflowTypes()
-		result = append(result, nextTypes...)
-	}
-	return result
-}
-
-func (r *registry) getActivityAlias(fnName string) (string, bool) {
-	r.Lock() // do not defer for Unlock to call next.getActivityAlias without lock
-	alias, ok := r.activityAliasMap[fnName]
-	if !ok && r.next != nil {
-		r.Unlock()
-		return r.next.getActivityAlias(fnName)
-	}
-	r.Unlock()
-	return alias, ok
-}
-
-// Use in unit test only, otherwise deadlock will occur.
-func (r *registry) addActivityWithLock(fnName string, a activity) {
-	r.Lock()
-	defer r.Unlock()
-	r.activityFuncMap[fnName] = a
-}
-
-func (r *registry) GetActivity(fnName string) (activity, bool) {
-	r.Lock() // do not defer for Unlock to call next.GetActivity without lock
-	a, ok := r.activityFuncMap[fnName]
-	if !ok && r.next != nil {
-		r.Unlock()
-		return r.next.GetActivity(fnName)
-	}
-	r.Unlock()
-	return a, ok
-}
-
-func (r *registry) getActivityNoLock(fnName string) (activity, bool) {
-	a, ok := r.activityFuncMap[fnName]
-	if !ok && r.next != nil {
-		return r.next.getActivityNoLock(fnName)
-	}
-	return a, ok
-}
-
-func (r *registry) getRegisteredActivities() []activity {
-	r.Lock() // do not defer for Unlock to call next.getRegisteredActivities without lock
-	activities := make([]activity, 0, len(r.activityFuncMap))
-	for _, a := range r.activityFuncMap {
-		activities = append(activities, a)
-	}
-	r.Unlock()
-	if r.next != nil {
-		nextActivities := r.next.getRegisteredActivities()
-		activities = append(activities, nextActivities...)
-	}
-	return activities
-}
-
-func isUseThriftEncoding(objs []interface{}) bool {
-	// NOTE: our criteria to use which encoder is simple if all the types are serializable using thrift then we use
-	// thrift encoder. For everything else we default to gob.
-
-	if len(objs) == 0 {
-		return false
-	}
-
-	for i := 0; i < len(objs); i++ {
-		if !isThriftType(objs[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func isUseThriftDecoding(objs []interface{}) bool {
-	// NOTE: our criteria to use which encoder is simple if all the types are de-serializable using thrift then we use
-	// thrift decoder. For everything else we default to gob.
-
-	if len(objs) == 0 {
-		return false
-	}
-
-	for i := 0; i < len(objs); i++ {
-		rVal := reflect.ValueOf(objs[i])
-		if rVal.Kind() != reflect.Ptr || !isThriftType(reflect.Indirect(rVal).Interface()) {
-			return false
-		}
-	}
-	return true
-}
-
-func (r *registry) getWorkflowDefinition(wt WorkflowType) (workflowDefinition, error) {
-	lookup := getFunctionName(wt.Name)
-	if alias, ok := r.getWorkflowAlias(lookup); ok {
-		lookup = alias
-	}
-	wf, ok := r.getWorkflowFn(lookup)
-	if !ok {
-		supported := strings.Join(r.getRegisteredWorkflowTypes(), ", ")
-		return nil, fmt.Errorf("unable to find workflow type: %v. Supported types: [%v]", lookup, supported)
-	}
-	wd := &workflowExecutor{workflowType: lookup, fn: wf}
-	return newSyncWorkflowDefinition(wd), nil
-}
-
 // Validate function parameters.
 func validateFnFormat(fnType reflect.Type, isWorkflow bool) error {
 	if fnType.Kind() != reflect.Func {
@@ -914,39 +642,6 @@ func decodeAndAssignValue(dc DataConverter, from interface{}, toValuePtr interfa
 	return nil
 }
 
-var typeOfByteSlice = reflect.TypeOf(([]byte)(nil))
-
-func isTypeByteSlice(inType reflect.Type) bool {
-	return inType == typeOfByteSlice || inType == reflect.PtrTo(typeOfByteSlice)
-}
-
-var once sync.Once
-
-// Singleton to hold the host registration details.
-var thImpl *registry
-
-func newRegistry() *registry {
-	return &registry{
-		workflowFuncMap:  make(map[string]interface{}),
-		workflowAliasMap: make(map[string]string),
-		activityFuncMap:  make(map[string]activity),
-		activityAliasMap: make(map[string]string),
-		next:             getGlobalRegistry(),
-	}
-}
-
-func getGlobalRegistry() *registry {
-	once.Do(func() {
-		thImpl = &registry{
-			workflowFuncMap:  make(map[string]interface{}),
-			workflowAliasMap: make(map[string]string),
-			activityFuncMap:  make(map[string]activity),
-			activityAliasMap: make(map[string]string),
-		}
-	})
-	return thImpl
-}
-
 // Wrapper to execute workflow functions.
 type workflowExecutor struct {
 	workflowType string
@@ -957,7 +652,7 @@ func (we *workflowExecutor) Execute(ctx Context, input []byte) ([]byte, error) {
 	var args []interface{}
 	dataConverter := getWorkflowEnvOptions(ctx).dataConverter
 	fnType := reflect.TypeOf(we.fn)
-	if fnType.NumIn() == 2 && isTypeByteSlice(fnType.In(1)) {
+	if fnType.NumIn() == 2 && util.IsTypeByteSlice(fnType.In(1)) {
 		// Do not deserialize input if workflow has a single byte slice argument (besides ctx)
 		args = append(args, input)
 	} else {
@@ -999,7 +694,7 @@ func (ae *activityExecutor) Execute(ctx context.Context, input []byte) ([]byte, 
 		args = append(args, reflect.ValueOf(ctx))
 	}
 
-	if fnType.NumIn() == 1 && isTypeByteSlice(fnType.In(0)) {
+	if fnType.NumIn() == 1 && util.IsTypeByteSlice(fnType.In(0)) {
 		args = append(args, reflect.ValueOf(input))
 	} else {
 		decoded, err := decodeArgs(dataConverter, fnType, input)
@@ -1215,231 +910,6 @@ func (aw *aggregatedWorker) Stop() {
 	aw.logger.Info("Stopped Worker")
 }
 
-// WorkflowReplayer is used to replay workflow code from an event history
-type WorkflowReplayer struct {
-	registry *registry
-}
-
-// NewWorkflowReplayer creates an instance of the WorkflowReplayer
-func NewWorkflowReplayer() *WorkflowReplayer {
-	return &WorkflowReplayer{registry: newRegistry()}
-}
-
-// RegisterWorkflow registers workflow function to replay
-func (aw *WorkflowReplayer) RegisterWorkflow(w interface{}) {
-	aw.registry.RegisterWorkflow(w)
-}
-
-// RegisterWorkflowWithOptions registers workflow function with custom workflow name to replay
-func (aw *WorkflowReplayer) RegisterWorkflowWithOptions(w interface{}, options RegisterWorkflowOptions) {
-	aw.registry.RegisterWorkflowWithOptions(w, options)
-}
-
-// ReplayWorkflowHistory executes a single decision task for the given history.
-// Use for testing backwards compatibility of code changes and troubleshooting workflows in a debugger.
-// The logger is an optional parameter. Defaults to the noop logger.
-func (aw *WorkflowReplayer) ReplayWorkflowHistory(logger *zap.Logger, history *shared.History) error {
-	if logger == nil {
-		logger = zap.NewNop()
-	}
-
-	testReporter := logger.Sugar()
-	controller := gomock.NewController(testReporter)
-	service := workflowservicetest.NewMockClient(controller)
-
-	return aw.replayWorkflowHistory(logger, service, ReplayDomainName, history)
-}
-
-// ReplayWorkflowHistoryFromJSONFile executes a single decision task for the given json history file.
-// Use for testing the backwards compatibility of code changes and troubleshooting workflows in a debugger.
-// The logger is an optional parameter. Defaults to the noop logger.
-func (aw *WorkflowReplayer) ReplayWorkflowHistoryFromJSONFile(logger *zap.Logger, jsonfileName string) error {
-	return aw.ReplayPartialWorkflowHistoryFromJSONFile(logger, jsonfileName, 0)
-}
-
-// ReplayPartialWorkflowHistoryFromJSONFile executes a single decision task for the given json history file upto provided
-// lastEventID(inclusive).
-// Use for testing backwards compatibility of code changes and troubleshooting workflows in a debugger.
-// The logger is an optional parameter. Defaults to the noop logger.
-func (aw *WorkflowReplayer) ReplayPartialWorkflowHistoryFromJSONFile(logger *zap.Logger, jsonfileName string, lastEventID int64) error {
-	history, err := extractHistoryFromFile(jsonfileName, lastEventID)
-
-	if err != nil {
-		return err
-	}
-
-	if logger == nil {
-		logger = zap.NewNop()
-	}
-
-	testReporter := logger.Sugar()
-	controller := gomock.NewController(testReporter)
-	service := workflowservicetest.NewMockClient(controller)
-
-	return aw.replayWorkflowHistory(logger, service, ReplayDomainName, history)
-}
-
-// ReplayWorkflowExecution replays workflow execution loading it from Temporal service.
-func (aw *WorkflowReplayer) ReplayWorkflowExecution(
-	ctx context.Context,
-	service workflowserviceclient.Interface,
-	logger *zap.Logger,
-	domain string,
-	execution WorkflowExecution,
-) error {
-	sharedExecution := &shared.WorkflowExecution{
-		RunId:      common.StringPtr(execution.RunID),
-		WorkflowId: common.StringPtr(execution.ID),
-	}
-	request := &shared.GetWorkflowExecutionHistoryRequest{
-		Domain:    common.StringPtr(domain),
-		Execution: sharedExecution,
-	}
-	hResponse, err := service.GetWorkflowExecutionHistory(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	if hResponse.RawHistory != nil {
-		history, err := serializer.DeserializeBlobDataToHistoryEvents(hResponse.RawHistory, shared.HistoryEventFilterTypeAllEvent)
-		if err != nil {
-			return err
-		}
-
-		hResponse.History = history
-	}
-
-	return aw.replayWorkflowHistory(logger, service, domain, hResponse.History)
-}
-
-func (aw *WorkflowReplayer) replayWorkflowHistory(
-	logger *zap.Logger,
-	service workflowserviceclient.Interface,
-	domain string,
-	history *shared.History,
-) error {
-	taskList := "ReplayTaskList"
-	events := history.Events
-	if events == nil {
-		return errors.New("empty events")
-	}
-	if len(events) < 3 {
-		return errors.New("at least 3 events expected in the history")
-	}
-	first := events[0]
-	if first.GetEventType() != shared.EventTypeWorkflowExecutionStarted {
-		return errors.New("first event is not WorkflowExecutionStarted")
-	}
-	last := events[len(events)-1]
-
-	attr := first.WorkflowExecutionStartedEventAttributes
-	if attr == nil {
-		return errors.New("corrupted WorkflowExecutionStarted")
-	}
-	workflowType := attr.WorkflowType
-	execution := &shared.WorkflowExecution{
-		RunId:      common.StringPtr(uuid.NewRandom().String()),
-		WorkflowId: common.StringPtr("ReplayId"),
-	}
-	if first.WorkflowExecutionStartedEventAttributes.GetOriginalExecutionRunId() != "" {
-		execution.RunId = common.StringPtr(first.WorkflowExecutionStartedEventAttributes.GetOriginalExecutionRunId())
-	}
-
-	task := &shared.PollForDecisionTaskResponse{
-		Attempt:                common.Int64Ptr(0),
-		TaskToken:              []byte("ReplayTaskToken"),
-		WorkflowType:           workflowType,
-		WorkflowExecution:      execution,
-		History:                history,
-		PreviousStartedEventId: common.Int64Ptr(math.MaxInt64),
-	}
-	if logger == nil {
-		logger = zap.NewNop()
-	}
-
-	metricScope := tally.NoopScope
-	iterator := &historyIteratorImpl{
-		nextPageToken: task.NextPageToken,
-		execution:     task.WorkflowExecution,
-		domain:        ReplayDomainName,
-		service:       service,
-		metricsScope:  metricScope,
-		maxEventID:    task.GetStartedEventId(),
-	}
-	params := workerExecutionParameters{
-		TaskList: taskList,
-		Identity: "replayID",
-		Logger:   logger,
-	}
-	taskHandler := newWorkflowTaskHandler(domain, params, nil, aw.registry)
-	resp, err := taskHandler.ProcessWorkflowTask(&workflowTask{task: task, historyIterator: iterator}, nil)
-	if err != nil {
-		return err
-	}
-
-	if last.GetEventType() != shared.EventTypeWorkflowExecutionCompleted && last.GetEventType() != shared.EventTypeWorkflowExecutionContinuedAsNew {
-		return nil
-	}
-	err = fmt.Errorf("replay workflow doesn't return the same result as the last event, resp: %v, last: %v", resp, last)
-	if resp != nil {
-		completeReq, ok := resp.(*shared.RespondDecisionTaskCompletedRequest)
-		if ok {
-			for _, d := range completeReq.Decisions {
-				if d.GetDecisionType() == shared.DecisionTypeContinueAsNewWorkflowExecution {
-					if last.GetEventType() == shared.EventTypeWorkflowExecutionContinuedAsNew {
-						inputA := d.ContinueAsNewWorkflowExecutionDecisionAttributes.Input
-						inputB := last.WorkflowExecutionContinuedAsNewEventAttributes.Input
-						if bytes.Compare(inputA, inputB) == 0 {
-							return nil
-						}
-					}
-				}
-				if d.GetDecisionType() == shared.DecisionTypeCompleteWorkflowExecution {
-					if last.GetEventType() == shared.EventTypeWorkflowExecutionCompleted {
-						resultA := last.WorkflowExecutionCompletedEventAttributes.Result
-						resultB := d.CompleteWorkflowExecutionDecisionAttributes.Result
-						if bytes.Compare(resultA, resultB) == 0 {
-							return nil
-						}
-					}
-				}
-			}
-		}
-	}
-	return err
-}
-
-func extractHistoryFromFile(jsonfileName string, lastEventID int64) (*shared.History, error) {
-	raw, err := ioutil.ReadFile(jsonfileName)
-	if err != nil {
-		return nil, err
-	}
-
-	var deserializedEvents []*shared.HistoryEvent
-	err = json.Unmarshal(raw, &deserializedEvents)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if lastEventID <= 0 {
-		return &shared.History{Events: deserializedEvents}, nil
-	}
-
-	// Caller is potentially asking for subset of history instead of all history events
-	events := []*shared.HistoryEvent{}
-	for _, event := range deserializedEvents {
-		events = append(events, event)
-		if event.GetEventId() == lastEventID {
-			// Copy history upto last event (inclusive)
-			break
-		}
-	}
-	history := &shared.History{Events: events}
-
-	return history, nil
-}
-
 // AggregatedWorker returns an instance to manage the workers. Use defaultConcurrentPollRoutineSize (which is 2) as
 // poller size. The typical RTT (round-trip time) is below 1ms within data center. And the poll API latency is about 5ms.
 // With 2 poller, we could achieve around 300~400 RPS.
@@ -1647,94 +1117,6 @@ func isInterfaceNil(i interface{}) bool {
 
 func getReadOnlyChannel(c chan struct{}) <-chan struct{} {
 	return c
-}
-
-// encoding is capable of encoding and decoding objects
-type encoding interface {
-	Marshal([]interface{}) ([]byte, error)
-	Unmarshal([]byte, []interface{}) error
-}
-
-// jsonEncoding encapsulates json encoding and decoding
-type jsonEncoding struct {
-}
-
-// Marshal encodes an array of object into bytes
-func (g jsonEncoding) Marshal(objs []interface{}) ([]byte, error) {
-	var buf bytes.Buffer
-	enc := json.NewEncoder(&buf)
-	for i, obj := range objs {
-		if err := enc.Encode(obj); err != nil {
-			if err == io.EOF {
-				return nil, fmt.Errorf("missing argument at index %d of type %T", i, obj)
-			}
-			return nil, fmt.Errorf(
-				"unable to encode argument: %d, %v, with json error: %v", i, reflect.TypeOf(obj), err)
-		}
-	}
-	return buf.Bytes(), nil
-}
-
-// Unmarshal decodes a byte array into the passed in objects
-func (g jsonEncoding) Unmarshal(data []byte, objs []interface{}) error {
-	dec := json.NewDecoder(bytes.NewBuffer(data))
-	for i, obj := range objs {
-		if err := dec.Decode(obj); err != nil {
-			return fmt.Errorf(
-				"unable to decode argument: %d, %v, with json error: %v", i, reflect.TypeOf(obj), err)
-		}
-	}
-	return nil
-}
-
-func isThriftType(v interface{}) bool {
-	// NOTE: Thrift serialization works only if the values are pointers.
-	// Thrift has a validation that it meets thift.TStruct which has Read/Write pointer receivers.
-
-	if reflect.ValueOf(v).Kind() != reflect.Ptr {
-		return false
-	}
-	t := reflect.TypeOf((*thrift.TStruct)(nil)).Elem()
-	return reflect.TypeOf(v).Implements(t)
-}
-
-// thriftEncoding encapsulates thrift serializer/de-serializer.
-type thriftEncoding struct{}
-
-// Marshal encodes an array of thrift into bytes
-func (g thriftEncoding) Marshal(objs []interface{}) ([]byte, error) {
-	tlist := []thrift.TStruct{}
-	for i := 0; i < len(objs); i++ {
-		if !isThriftType(objs[i]) {
-			return nil, fmt.Errorf("pointer to thrift.TStruct type is required for %v argument", i+1)
-		}
-		t := reflect.ValueOf(objs[i]).Interface().(thrift.TStruct)
-		tlist = append(tlist, t)
-	}
-	return common.TListSerialize(tlist)
-}
-
-// Unmarshal decodes an array of thrift into bytes
-func (g thriftEncoding) Unmarshal(data []byte, objs []interface{}) error {
-	tlist := []thrift.TStruct{}
-	for i := 0; i < len(objs); i++ {
-		rVal := reflect.ValueOf(objs[i])
-		if rVal.Kind() != reflect.Ptr || !isThriftType(reflect.Indirect(rVal).Interface()) {
-			return fmt.Errorf("pointer to pointer thrift.TStruct type is required for %v argument", i+1)
-		}
-		t := reflect.New(rVal.Elem().Type().Elem()).Interface().(thrift.TStruct)
-		tlist = append(tlist, t)
-	}
-
-	if err := common.TListDeserialize(tlist, data); err != nil {
-		return err
-	}
-
-	for i := 0; i < len(tlist); i++ {
-		reflect.ValueOf(objs[i]).Elem().Set(reflect.ValueOf(tlist[i]))
-	}
-
-	return nil
 }
 
 func augmentWorkerOptions(options WorkerOptions) WorkerOptions {
